# indexer/swagger.yaml
swagger: "2.0"
info:
  title: Mark3d service
  description: >
    Здесь можно **маркдауном** писать дополнительное описание.<br />
  version: 1.0.0
schemes:
  - https
host: filemarket.xyz
basePath: /api
paths:
...
  /collections/full/public:
    get:
      tags:
        - collections
      description: Получение публичной коллекции со всем ее токенами
      produces:
        - application/json
      responses:
        "200":
          description: Коллекция
          schema:
            $ref: "#/definitions/CollectionData"
        default:
          description: Ошибка
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /orders/all_active:
    get:
      tags:
        - orders
      description: Получение списка всех активных ордеров с информацией о токенах
      produces:
        - application/json
      responses:
        200:
          description: Список ордеров
          schema:
            type: array
            items:
              $ref: "#/definitions/OrderWithToken"
        default:
          description: Ошибка
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /sequencer/acquire:
    get:
      description: Возвращает рандомный доступный tokenId
      produces:
        - application/json
      responses:
        "200":
          description: Token Id
          schema:
            $ref: "#/definitions/SequencerAcquireResponse"
        default:
          description: Ошибка
          schema:
            $ref: '#/definitions/ErrorResponse'
definitions:
  ErrorResponse:
    type: object
    description: Типовая ошибка
    properties:
      code:
        type: integer
        description: Код ошибки
      message:
        type: string
        description: Сообщение ошибки
      detail:
        type: string
        description: Дополнительные детали
...
  OrderWithToken:
    type: object
    description: Ордер с токеном
    properties:
      order:
        $ref: "#/definitions/Order"
      token:
        $ref: "#/definitions/Token"
      transfer:
        $ref: "#/definitions/Transfer"
  CollectionData:
    type: object
    description: Коллекция со всеми токенами
    properties:
      collection:
        $ref: "#/definitions/Collection"
      tokens:
        type: array
        items:
          $ref: "#/definitions/Token"
  Collection:
    type: object
    description: Коллекция
    properties:
      address:
        type: string
        description: Адрес контракта коллекции
        pattern: "0x[0-9a-f-A-F]{40}"
      creator:
        type: string
        description: Создатель коллекции
        pattern: "0x[0-9a-f-A-F]{40}"
      owner:
        type: string
        description: Владелец коллекции
        pattern: "0x[0-9a-f-A-F]{40}"
      tokenId:
        type: string
        description: Id Access токена
        pattern: "^[0-9]+$"
      metaUri:
        type: string
        description: uri метаданных access токена
      name:
        type: string
        description: Имя токена из метаданных
      description:
        type: string
        description: Описание токена из метаданных
      image:
        type: string
        description: Ссылка на пикчу из метаданных
      type:
        type: string
        description: Тип коллекции <br />
          * `File Market Collection` - Обычные коллекции <br />
          * `Public Collection` - Общая коллекция, куда может заминтить каждый <br />
          * `File Bunnies Collection` - Крош. Наша коллекция
        enum:
          - "File Market Collection"
          - "Public Collection"
          - "File Bunnies Collection"
  Token:
    type: object
    description: Токен
    properties:
      collectionAddress:
        type: string
        description: Адрес коллекции
        pattern: "0x[0-9a-f-A-F]{40}"
      collectionName:
        type: string
        description: Имя коллекции
      tokenId:
        type: string
        description: Id токена
        pattern: "^[0-9]+$"
      owner:
        type: string
        description: Владелец
        pattern: "0x[0-9a-f-A-F]{40}"
      creator:
        type: string
        description: Создатель
        pattern: "0x[0-9a-f-A-F]{40}"
      metaUri:
        type: string
        description: uri метаданных токена
      name:
        type: string
        description: Метаданные - Имя токена из метаданных
      description:
        type: string
        description: Метаданные - Описание токена из метаданных
      image:
        type: string
        description: Метаданные - Ссылка на пикчу из метаданных
      externalLink:
        type: string
        description: не интернал не амнистия
      hiddenFile:
        type: string
        description: Метаданные - Ссылка на зашифрованный файл из метаданных
      license:
        type: string
      licenseUrl:
        type: string
      royalty:
        type: number
        description: Роялти в формате, где 100% = 10000
        format: uint64
      mintTxHash:
        type: string
        description: Хэш транзакции минта
      mintTxTimestamp:
        type: number
        description: Время транзакции минта
        format: uint64
      hiddenFileMeta:
        $ref: "#/definitions/HiddenFileMetaData"
      categories:
        type: array
        description: Метаданные - Категории
        items:
          type: string
      subcategories:
        type: array
        description: Метаданные - Подкатегории
        items:
          type: string
      tags:
        type: array
        items:
          type: string
        description: Метаданные - Тэги
      properties:
        type: array
        items:
          $ref: "#/definitions/MetadataProperty"
        description: Метаданные - Аттрибуты
      rankings:
        type: array
        items:
          $ref: "#/definitions/MetadataProperty"
        description: Метаданные - Ранкинги
      stats:
        type: array
        items:
          $ref: "#/definitions/MetadataProperty"
        description: Метаданные - Статы
      pendingTransferId:
        type: integer
        format: int64
        description: id трансфера, находящегося в процессе
      pendingOrderId:
        type: integer
        format: int64
        description: id ордера, находящегося в процессе
...
  Transfer:
    type: object
    description: Трансфер
    properties:
      id:
        type: integer
        description: Id трансфера
        format: int64
      collection:
        type: string
        description: Адрес коллекции
        pattern: "0x[0-9a-f-A-F]{40}"
      tokenId:
        type: string
        description: Id токена
        pattern: "^[0-9]+$"
      from:
        type: string
        description: Отправитель
        pattern: "0x[0-9a-f-A-F]{40}"
      to:
        type: string
        description: Получатель, если уже назначен
        pattern: "0x[0-9a-f-A-F]{40}"
      orderId:
        type: integer
        description: Id ордера, если трансфер создан из-за ордера
        format: int64
      fraudApproved:
        type: boolean
        description:
          true - если факт мошенничества подтвержден (репорт подает всегда только получатель).
          был ли подан репорт, можно выяснить по статусам
      publicKey:
        type: string
        description: публичный ключ (hex)
      encryptedPassword:
        type: string
        description: зашифрованный пароль (hex)
      statuses:
        type: array
        description: Статусы трансфера
        items:
          $ref: "#/definitions/TransferStatusInfo"
      number:
        type: string
        description: Порядковый номер трансфера
  Order:
    type: object
    description: Ордер
    properties:
      id:
        type: integer
        description: Id ордера
        format: int64
      transferId:
        type: integer
        description: Id трансфера
        format: int64
      price:
        type: string
        description: Цена
        pattern: "^[0-9]+$"
      currency:
        type: string
        description: Адрес ERC20. Если нулевой адрес, то нативная валюта
        pattern: "0x[0-9a-f-A-F]{40}"
      exchangeAddress:
        type: string
        description: Адрес коллекции
        pattern: "0x[0-9a-f-A-F]{40}"
      statuses:
        type: array
        description: Статусы ордера
        items:
          $ref: "#/definitions/OrderStatusInfo"
  TransferStatusInfo:
    type: object
    description: Изменение статуса трансфера
    properties:
      status:
        $ref: "#/definitions/TransferStatus"
      timestamp:
        type: number
        description: Время установки статуса
        format: int64
      txId:
        type: string
        description: Транзакция, которой был установлен статус
        pattern: "0x[0-9a-f-A-F]{64}"
  TransferStatus:
    description: Статус трансфера <br />
      * `Drafted` - Создан черновик трансфера <br />
      * `Created` - Трансфер инициализирован <br />
      * `PublicKeySet` - Сохранен публичный ключ получателя <br />
      * `PasswordSet` - Сохранен зашифрованный пароль <br />
      * `FraudReported` - Доложено мошенничество <br />
      * `Finished` - Трансфер завершен <br />
      * `Cancelled` - Трансфер отменен
    type: string
    enum:
      - Drafted
      - Created
      - PublicKeySet
      - PasswordSet
      - FraudReported
      - Finished
      - Cancelled
  OrderStatusInfo:
    type: object
    description: Изменение статуса ордера
    properties:
      status:
        $ref: "#/definitions/OrderStatus"
      timestamp:
        type: number
        description: Время установки статуса
        format: int64
      txId:
        type: string
        description: Транзакция, которой был установлен статус
        pattern: "0x[0-9a-f-A-F]{64}"
  OrderStatus:
    description: Статус ордера <br />
      * `Created` - ордер создан <br />
      * `Fulfilled` - средства внесены покупателем <br />
      * `Finished` - ордер завершен успешно <br />
      * `FraudApproved` - доказано мошенничество и ордер отменен <br />
      * `Cancelled` - ордер отменен
    type: string
    enum:
      - Created
      - Fulfilled
      - Finished
      - FraudApproved
      - Cancelled
...
  SequencerAcquireResponse:
    type: object
    properties:
      tokenId:
        type: string
        description: Id токена
        pattern: "^[0-9]+$"
  MetadataProperty:
    type: object
    description: Структура для свойств в метаданных
    properties:
      traitType:
        type: string
        description: Название свойства
      displayType:
        type: string
        description: Тип данных значения
      value:
        type: string
        description: Значение свойства
      max_value:
        type: string
      min_value:
        type: string
      traitValueCount:
        type: integer
        description: Количество таких же value в заданном trait_type
      traitTotal:
        type: integer
        description: Общее количество записей с таким trait_type
  HiddenFileMetaData:
    type: object
    description: Метаданные скрытого файла
    properties:
      name:
        type: string
        description: Имя файла
      type:
        type: string
        description: Тип файла
      size:
        type: number
        format: int64
        description: Размер данных
...